Feat : 경매물물수정삭제 20230913 1409

See merge request s09-fintech-finance-sub2/S09P22B207!35

+++ Front-End/pro/app/src/main/java/com/example/project/viewmodels/BiometricViewModel.kt
@@ -22,7 +22,7 @@ class BiometricViewModel @Inject constructor(
 
     fun verifyPattern(pattern: String) {
         viewModelScope.launch {
-            val userId = sharedPreferencesUtil.getUserId().toLong()
+            val userId = sharedPreferencesUtil.getUserId()
             if (userId == null) {
                 _authenticationState.value = AuthenticationState.ERROR("정상적인 경로로 회원가입을 진행후 로그인 해주세요.")
                 return@launch
